MTP and Exodus bugfix release (#308)
* Change all Exodus console output to log file instead

* Added MTP bit to consensus parameter

* Fixed crash when reindexing with unconfirmed spends in wallet

* MTP Malformed proof unit tests

* MTP tests update

* Disable Exodus by default

* Rename Balances page to ExoAssets and move to before Toolbox

* Saving znodes & znode payments cache

* Allow mining only if znode winner is known

* Fix incorrect tooltip for Send page

* Fixed potential deadlock

* Version bump to 0.13.7.2

* Update Exodus option's usage and remove unused one

* Remove Exodus initial scan block filter since we don't use it

+++ src/init.cpp
@@ -17,6 +17,7 @@
 #include "checkpoints.h"
 #include "compat/sanity.h"
 #include "consensus/validation.h"
+#include "exodus/exodus.h"
 #include "httpserver.h"
 #include "httprpc.h"
 #include "key.h"
@@ -119,10 +120,6 @@ enum BindFlags {
 static const char *FEE_ESTIMATES_FILENAME = "fee_estimates.dat";
 
 extern CTxMemPool stempool;
-// Exodus initialization and shutdown handlers
-extern int exodus_init();
-extern int exodus_shutdown();
-extern int CheckWalletUpdate(bool forceUpdate = false);
 
 namespace fs = boost::filesystem;
 
@@ -248,14 +245,12 @@ void Shutdown() {
     GenerateBitcoins(false, 0, Params());
     StopNode();
 
-    // STORE DATA CACHES INTO SERIALIZED DAT FILES
-    // TODO: https://github.com/zcoinofficial/zcoin/issues/182
-    //    CFlatDB<CZnodeMan> flatdb1("zncache.dat", "magicZnodeCache");
-    //    flatdb1.Dump(mnodeman);
-    //    CFlatDB<CZnodePayments> flatdb2("znpayments.dat", "magicZnodePaymentsCache");
-    //    flatdb2.Dump(mnpayments);
-    //    CFlatDB<CNetFulfilledRequestManager> flatdb4("netfulfilled.dat", "magicFulfilledCache");
-    //    flatdb4.Dump(netfulfilledman);
+    CFlatDB<CZnodeMan> flatdb1("zncache.dat", "magicZnodeCache");
+    flatdb1.Dump(mnodeman);
+    CFlatDB<CZnodePayments> flatdb2("znpayments.dat", "magicZnodePaymentsCache");
+    flatdb2.Dump(mnpayments);
+    CFlatDB<CNetFulfilledRequestManager> flatdb4("netfulfilled.dat", "magicFulfilledCache");
+    flatdb4.Dump(netfulfilledman);
 
     StopTorControl();
     UnregisterNodeSignals(GetNodeSignals());
@@ -285,8 +280,9 @@ void Shutdown() {
         pblocktree = NULL;
     }
 
-    //! Exodus shutdown
-    exodus_shutdown();
+    if (isExodusEnabled()) {
+        exodus_shutdown();
+    }
 
 #ifdef ENABLE_WALLET
     if (pwalletMain)
@@ -669,21 +665,19 @@ std::string HelpMessage(HelpMessageMode mode) {
     }
 
     strUsage += HelpMessageGroup("Exodus options:");
-	strUsage += HelpMessageOpt("-startclean", "Clear all persistence files on startup; triggers reparsing of Exodus transactions (default: 0)");
-	strUsage += HelpMessageOpt("-exodustxcache", "The maximum number of transactions in the input transaction cache (default: 500000)");
-	strUsage += HelpMessageOpt("-exodusprogressfrequency", "Time in seconds after which the initial scanning progress is reported (default: 30)");
-	strUsage += HelpMessageOpt("-exodusseedblockfilter", "Set skipping of blocks without Exodus transactions during initial scan (default: 1)");
-	strUsage += HelpMessageOpt("-exoduslogfile", "The path of the log file (default: exodus.log)");
-	strUsage += HelpMessageOpt("-exodusdebug=<category>", "Enable or disable log categories, can be \"all\" or \"none\"");
-	strUsage += HelpMessageOpt("-autocommit", "Enable or disable broadcasting of transactions, when creating transactions (default: 1)");
-	strUsage += HelpMessageOpt("-overrideforcedshutdown", "Overwrite shutdown, triggered by an alert (default: 0)");
-	strUsage += HelpMessageOpt("-exodusalertallowsender", "Whitelist senders of alerts, can be \"any\")");
-	strUsage += HelpMessageOpt("-exodusalertignoresender", "Ignore senders of alerts");
-	strUsage += HelpMessageOpt("-exodusactivationignoresender", "Ignore senders of activations");
-	strUsage += HelpMessageOpt("-exodusactivationallowsender", "Whitelist senders of activations");
-	strUsage += HelpMessageOpt("-disclaimer", "Explicitly show QT disclaimer on startup (default: 0)");
-	strUsage += HelpMessageOpt("-exodusuiwalletscope", "Max. transactions to show in trade and transaction history (default: 65535)");
-	strUsage += HelpMessageOpt("-exodusshowblockconsensushash", "Calculate and log the consensus hash for the specified block");
+    strUsage += HelpMessageOpt("-exodus", "Enable Exodus");
+    strUsage += HelpMessageOpt("-startclean", "Clear all persistence files on startup; triggers reparsing of Exodus transactions");
+    strUsage += HelpMessageOpt("-exodustxcache=<num>", "The maximum number of transactions in the input transaction cache (default: 500000)");
+    strUsage += HelpMessageOpt("-exodusprogressfrequency=<seconds>", "Time in seconds after which the initial scanning progress is reported (default: 30)");
+    strUsage += HelpMessageOpt("-exodusdebug=<category>", "Enable or disable log categories, can be \"all\" or \"none\"");
+    strUsage += HelpMessageOpt("-autocommit=<flag>", "Enable or disable broadcasting of transactions, when creating transactions (default: 1)");
+    strUsage += HelpMessageOpt("-overrideforcedshutdown=<flag>", "Disable force shutdown when error (default: 0)");
+    strUsage += HelpMessageOpt("-exodusalertallowsender=<addr>", "Whitelist senders of alerts, can be \"any\")");
+    strUsage += HelpMessageOpt("-exodusalertignoresender=<addr>", "Ignore senders of alerts");
+    strUsage += HelpMessageOpt("-exodusactivationignoresender=<addr>", "Ignore senders of activations");
+    strUsage += HelpMessageOpt("-exodusactivationallowsender=<addr>", "Whitelist senders of activations");
+    strUsage += HelpMessageOpt("-exodusuiwalletscope=<number>", "Max. transactions to show in trade and transaction history (default: 65535)");
+    strUsage += HelpMessageOpt("-exodusshowblockconsensushash=<number>", "Calculate and log the consensus hash for the specified block");
 
     return strUsage;
 }
@@ -1641,7 +1635,7 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
                 delete pblocktree;
 
                 pblocktree = new CBlockTreeDB(nBlockTreeDBCache, false, fReindex);
-	            
+
 	            if (!fReindex) {
                     // Check existing block index database version, reindex if needed
                     if (pblocktree->GetBlockIndexVersion() < ZC_ADVANCED_INDEX_VERSION) {
@@ -1651,7 +1645,7 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
                         pblocktree = new CBlockTreeDB(nBlockTreeDBCache, false, fReindex);
 		            }
 	            }
-	            
+
                 pcoinsdbview = new CCoinsViewDB(nCoinDBCache, false, fReindex || fReindexChainState);
                 pcoinscatcher = new CCoinsViewErrorCatcher(pcoinsdbview);
                 pcoinsTip = new CCoinsViewCache(pcoinscatcher);
@@ -1771,44 +1765,45 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
 
     // ********************************************************* Step 7.5: load exodus
 
-	if (!fTxIndex) {
-		// ask the user if they would like us to modify their config file for them
-		std::string msg = _("Disabled transaction index detected.\n\n"
-							"Exodus requires an enabled transaction index. To enable "
-							"transaction indexing, please use the \"-txindex\" option as "
-							"command line argument or add \"txindex=1\" to your client "
-							"configuration file within your data directory.\n\n"
-							"Configuration file"); // allow translation of main text body while still allowing differing config file string
-		msg += ": " + GetConfigFile().string() + "\n\n";
-		msg += _("Would you like Exodus to attempt to update your configuration file accordingly?");
-		bool fRet = uiInterface.ThreadSafeMessageBox(msg, "", CClientUIInterface::MSG_INFORMATION | CClientUIInterface::BTN_OK | CClientUIInterface::MODAL | CClientUIInterface::BTN_ABORT);
-		if (fRet) {
-			// add txindex=1 to config file in GetConfigFile()
-			boost::filesystem::path configPathInfo = GetConfigFile();
-			FILE *fp = fopen(configPathInfo.string().c_str(), "at");
-			if (!fp) {
-				std::string failMsg = _("Unable to update configuration file at");
-				failMsg += ":\n" + GetConfigFile().string() + "\n\n";
-				failMsg += _("The file may be write protected or you may not have the required permissions to edit it.\n");
-				failMsg += _("Please add txindex=1 to your configuration file manually.\n\nExodus will now shutdown.");
-				return InitError(failMsg);
-			}
-			fprintf(fp, "\ntxindex=1\n");
-			fflush(fp);
-			fclose(fp);
-			std::string strUpdated = _(
-					"Your configuration file has been updated.\n\n"
-					"Exodus will now shutdown - please restart the client for your new configuration to take effect.");
-			uiInterface.ThreadSafeMessageBox(strUpdated, "", CClientUIInterface::MSG_INFORMATION | CClientUIInterface::BTN_OK | CClientUIInterface::MODAL);
-			return false;
-		} else {
-			return InitError(_("Please add txindex=1 to your configuration file manually.\n\nOmni Core will now shutdown."));
-		}
-	}
+    if (isExodusEnabled()) {
+        if (!fTxIndex) {
+            // ask the user if they would like us to modify their config file for them
+            std::string msg = _("Disabled transaction index detected.\n\n"
+                                "Exodus requires an enabled transaction index. To enable "
+                                "transaction indexing, please use the \"-txindex\" option as "
+                                "command line argument or add \"txindex=1\" to your client "
+                                "configuration file within your data directory.\n\n"
+                                "Configuration file"); // allow translation of main text body while still allowing differing config file string
+            msg += ": " + GetConfigFile().string() + "\n\n";
+            msg += _("Would you like Exodus to attempt to update your configuration file accordingly?");
+            bool fRet = uiInterface.ThreadSafeMessageBox(msg, "", CClientUIInterface::MSG_INFORMATION | CClientUIInterface::BTN_OK | CClientUIInterface::MODAL | CClientUIInterface::BTN_ABORT);
+            if (fRet) {
+                // add txindex=1 to config file in GetConfigFile()
+                boost::filesystem::path configPathInfo = GetConfigFile();
+                FILE *fp = fopen(configPathInfo.string().c_str(), "at");
+                if (!fp) {
+                    std::string failMsg = _("Unable to update configuration file at");
+                    failMsg += ":\n" + GetConfigFile().string() + "\n\n";
+                    failMsg += _("The file may be write protected or you may not have the required permissions to edit it.\n");
+                    failMsg += _("Please add txindex=1 to your configuration file manually.\n\nExodus will now shutdown.");
+                    return InitError(failMsg);
+                }
+                fprintf(fp, "\ntxindex=1\n");
+                fflush(fp);
+                fclose(fp);
+                std::string strUpdated = _(
+                        "Your configuration file has been updated.\n\n"
+                        "Exodus will now shutdown - please restart the client for your new configuration to take effect.");
+                uiInterface.ThreadSafeMessageBox(strUpdated, "", CClientUIInterface::MSG_INFORMATION | CClientUIInterface::BTN_OK | CClientUIInterface::MODAL);
+                return false;
+            } else {
+                return InitError(_("Please add txindex=1 to your configuration file manually.\n\nOmni Core will now shutdown."));
+            }
+        }
 
-	uiInterface.InitMessage(_("Parsing Exodus transactions..."));
-
-	exodus_init();
+        uiInterface.InitMessage(_("Parsing Exodus transactions..."));
+        exodus_init();
+    }
 
     // ********************************************************* Step 8: load wallet
 
@@ -1826,8 +1821,10 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
     LogPrintf("No wallet support compiled in!\n");
 #endif // !ENABLE_WALLET
 
-    // Exodus code should be initialized and wallet should now be loaded, perform an initial populat$
-    CheckWalletUpdate();
+    // Exodus code should be initialized and wallet should now be loaded, perform an initial populate
+    if (isExodusEnabled()) {
+        CheckWalletUpdate();
+    }
 
     // ********************************************************* Step 9: data directory maintenance
     LogPrintf("Step 9: data directory maintenance **********************\n");
@@ -2001,8 +1998,7 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
     // ********************************************************* Step 11b: Load cache data
 
     // LOAD SERIALIZED DAT FILES INTO DATA CACHES FOR INTERNAL USE
-    // TODO: https://github.com/zcoinofficial/zcoin/issues/182
-    /* uiInterface.InitMessage(_("Loading znode cache..."));
+    uiInterface.InitMessage(_("Loading znode cache..."));
     CFlatDB<CZnodeMan> flatdb1("zncache.dat", "magicZnodeCache");
     if (!flatdb1.Load(mnodeman)) {
         return InitError("Failed to load znode cache from zncache.dat");
@@ -2015,13 +2011,13 @@ bool AppInit2(boost::thread_group &threadGroup, CScheduler &scheduler) {
             return InitError("Failed to load znode payments cache from znpayments.dat");
         }
     } else {
-        uiInterface.InitMessage(_("Znode cache is empty, skipping payments and governance cache..."));
-    } */
+        uiInterface.InitMessage(_("Znode cache is empty, skipping payments cache..."));
+    }
 
-    // uiInterface.InitMessage(_("Loading fulfilled requests cache..."));
+    uiInterface.InitMessage(_("Loading fulfilled requests cache..."));
     CFlatDB<CNetFulfilledRequestManager> flatdb4("netfulfilled.dat", "magicFulfilledCache");
 	flatdb4.Load(netfulfilledman);
-	
+
     // if (!flatdb4.Load(netfulfilledman)) {
     //     LogPrint"Failed to load fulfilled requests cache from netfulfilled.dat");
     // }
